fix: repair project 040 runtime and parking APIs

+++ 040-backend/src/main/java/com/community/parking/dto/ReportDTO.java
@@ -2,13 +2,15 @@ package com.community.parking.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class ReportDTO {
     private String plateNumber;
     private Long violationTypeId;
     private String location;
-    private Double longitude;
-    private Double latitude;
+    private BigDecimal longitude;
+    private BigDecimal latitude;
     private String description;
     private String images;
     private Boolean isAnonymous;
